fix types

+++ FlockingBoids/Model/Position.cs
@@ -23,15 +23,15 @@ namespace Model
             Y += position.Y;
         }
 
-        public (double x, double y) Delta(Position otherPosition)
+        public (float x, float y) Delta(Position otherPosition)
         {
             return (otherPosition.X - X, otherPosition.Y - Y);
         }
 
-        public double Distance(Position otherPosition)
+        public float Distance(Position otherPosition)
         {
             var (dX, dY) = Delta(otherPosition);
-            return Math.Sqrt(dX * dX + dY * dY);
+            return (float)Math.Sqrt(dX * dX + dY * dY);
         }
 
     }
